refactor: relocate api and pages
