+++ frontend/src/context/ApplicationProvider.tsx
@@ -5,8 +5,8 @@ import {
   deleteAttachment as apiDeleteAttachment,
   updateApplication,
   getApplicationAttachments,
-} from "../lib/api/applications";
-import { getCall } from "../lib/api/calls";
+} from "../api/applications";
+import { getCall } from "../api/calls";
 import { Call } from "../types";
 import { Attachment } from "../types/application";
